requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
faiss-cpu
python-dotenv
sentence-transformers
